Initial commit
This is the initial version of the session-select.nvim plugin. I
extracted it from my Neovim config, wrote a README for it, and turned it
into a standalone plugin. I don't know if anyone else will use it, but I
think that it's gotten large enough that it deserves to be maintained as
its own thing regardless.